Update inference_model_with_image (DIPA 1.0).py

+++ dataset/inference_model_with_image (DIPA 1.0).py
@@ -63,7 +63,7 @@ if __name__ == '__main__':
     val_df = mega_table.drop(train_df.index)
 
     train_df = Crowdworks_data
-    val_df = prolific_data.sample(n = 0.2 * len(train_df), random_state=0)
+    val_df = prolific_data.sample(n = int(0.2 * len(train_df)), random_state=0)
     train_dataset = ImageMaskDataset(train_df, image_folder, label_folder, input_channel, output_name, image_size, flip = True)
     val_dataset = ImageMaskDataset(val_df, image_folder, label_folder, input_channel, output_name, image_size)
